pyproject.toml
[project]
name = "base-template-py"
version = "0.1.0"
description = ""
authors = [{ name = "Hugo Lacerda", email = "hugolacerdar@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "freezegun (>=1.5.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "tenacity (>=9.0.0,<10.0.0)",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-watch = "^4.2.0"
coverage = "^7.6.10"
pytest-asyncio = "^0.25.2"
ruff = "^0.9.1"
pyright = "^1.1.391"
detect-secrets = "^1.5.0"
pre-commit = "^4.0.1"
commitizen = "^4.1.0"
gitlint = "^0.19.1"

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
pythonpath = ["."]
minversion = "6.0"
addopts = ["-vv", "--strict-config", "--strict-markers"]
testpaths = ["tests"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope="function"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["*__init__.py", "*abc.py", "*environment.py", "*app.py", "*app_builder.py", "*settings.py", "*dependencies.py"]

[tool.ruff]
line-length = 140
lint.select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # fastapi
    "FAST",
    # annotations
    "ANN",
    # bandit
    "S",
]
lint.ignore = ["B008"]
lint.extend-safe-fixes = ["SIM", "I"]
format.quote-style = "single"
format.indent-style = "tab"
format.docstring-code-format = true

[tool.ruff.lint.per-file-ignores]
'tests/*' = ['S101', 'S104', 'S105', 'S311', 'ANN']

[tool.pyright]
typeCheckingMode = "strict"

[build-system]
requires      = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
